+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 # [1.4.14] - Unreleased
 - Add InputExt::cursor_color() and set_cursor_color() by @thorio.
 - Add TextBuffer::redo() and TextEditor::redo().
+- Add TextBuffer::append2().
 - Add TextBuffer::set_bytes().
 - Load system icons on linux.
 - Update FLTK submodule.